+++ simpleWebservice/src/main/java/com/kata/SimpleWebservice/User.java
@@ -1,33 +1,19 @@
 package com.kata.SimpleWebservice;
 
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import java.time.LocalDate;
+import lombok.Getter;
 
-@Setter
 @Getter
-@Entity
-@NoArgsConstructor
-@Table(name = "example_user")
-public class User {
-    @JsonIgnore
-    private @Id long id;
-    private String name;
-    private int age;
-    @Column(name="dateofbirth")
-    private LocalDate dateOfBirth;
+public final class User {
+	private final long id;
+	private final String name;
+	private final int age;
+	private final LocalDate dateOfBirth;
 
-    public User(String name, int age, String dateOfBirth){
-        this.name = name;
-        this.age = age;
-        this.dateOfBirth = LocalDate.parse(dateOfBirth);
-    }
+	public User(long id, String name, int age, LocalDate dateOfBirth) {
+		this.id = id;
+		this.name = name;
+		this.age = age;
+		this.dateOfBirth = dateOfBirth;
+	}
 }
