+++ simpleWebservice/src/test/java/com/kata/SimpleWebservice/UserTest.java
@@ -2,6 +2,7 @@ package com.kata.SimpleWebservice;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.time.LocalDate;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
@@ -16,9 +17,9 @@ public class UserTest {
 
     @Test
     void userObjectDeserializesCorrectly() throws JsonProcessingException {
-        User user = new User("John Doe", 26, "1970-12-31");
+        User user = new User(0,"John Doe", 26, LocalDate.of(1970,12,31));
         String actualJson = objectMapper.writeValueAsString(user);
-        String expectedJson = "{\"name\":\"John Doe\",\"age\":26,\"dateOfBirth\":\"1970-12-31\"}";
+        String expectedJson = "{\"id\":0,\"name\":\"John Doe\",\"age\":26,\"dateOfBirth\":\"1970-12-31\"}";
 
         assertThat(actualJson).isEqualTo(expectedJson);
     }
